Minor refactor to prevent circular dependencies. Linting.
