+++ contentcuration/contentcuration/frontend/shared/data/index.js
@@ -4,12 +4,12 @@ import { createLeaderElection } from './leaderElection';
 import channel from './broadcastChannel';
 import { CHANGE_LOCKS_TABLE, CHANGE_TYPES, CHANGES_TABLE } from './constants';
 import db, { CLIENTID } from './db';
-import { INDEXEDDB_RESOURCES } from './resources';
+import { INDEXEDDB_RESOURCES } from './registry';
 import { startSyncing, stopSyncing } from './serverSync';
 
 // Re-export for ease of reference.
-export { CHANGE_TYPES } from './constants';
-export { TABLE_NAMES, API_RESOURCES, INDEXEDDB_RESOURCES } from './resources';
+export { CHANGE_TYPES, TABLE_NAMES } from './constants';
+export { API_RESOURCES, INDEXEDDB_RESOURCES } from './registry';
 
 const LISTENERS = {};
