+++ contentcuration/contentcuration/frontend/shared/data/registry.js
@@ -1,0 +1,4 @@
+// Put resource registry objects in here to avoid circular import issues
+export const API_RESOURCES = {};
+
+export const INDEXEDDB_RESOURCES = {};
